+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/IssueCredentialProtocol.kt
@@ -7,6 +7,7 @@ import at.asitplus.wallet.lib.data.SchemaIndex
 import at.asitplus.wallet.lib.data.dif.CredentialDefinition
 import at.asitplus.wallet.lib.data.dif.CredentialManifest
 import at.asitplus.wallet.lib.data.jsonSerializer
+import at.asitplus.wallet.lib.jws.JsonWebKey
 import at.asitplus.wallet.lib.msg.AttachmentFormatReference
 import at.asitplus.wallet.lib.msg.IssueCredential
 import at.asitplus.wallet.lib.msg.IssueCredentialBody
@@ -100,13 +101,13 @@ class IssueCredentialProtocol(
         return createRequestCredential()
     }
 
-    override suspend fun parseMessage(body: JsonWebMessage, senderKeyId: String): InternalNextMessage {
+    override suspend fun parseMessage(body: JsonWebMessage, senderKey: JsonWebKey): InternalNextMessage {
         when (this.state) {
             State.START -> {
                 if (body is OutOfBandInvitation)
-                    return createRequestCredential(body, senderKeyId)
+                    return createRequestCredential(body, senderKey)
                 if (body is RequestCredential)
-                    return issueCredential(body, senderKeyId)
+                    return issueCredential(body, senderKey)
                 return InternalNextMessage.IncorrectState("messageType")
                     .also { Napier.w("Unexpected messageType: ${body.type}") }
             }
@@ -118,7 +119,7 @@ class IssueCredentialProtocol(
                 if (body.parentThreadId != invitationId)
                     return InternalNextMessage.IncorrectState("parentThreadId")
                         .also { Napier.w("Unexpected parentThreadId: ${body.parentThreadId}") }
-                return issueCredential(body, senderKeyId)
+                return issueCredential(body, senderKey)
             }
 
             State.REQUEST_CREDENTIAL_SENT -> {
@@ -166,7 +167,7 @@ class IssueCredentialProtocol(
             .also { this.state = State.REQUEST_CREDENTIAL_SENT }
     }
 
-    private fun createRequestCredential(invitation: OutOfBandInvitation, senderKeyId: String): InternalNextMessage {
+    private fun createRequestCredential(invitation: OutOfBandInvitation, senderKey: JsonWebKey): InternalNextMessage {
         val credentialScheme = ConstantIndex.Parser.parseGoalCode(invitation.body.goalCode)
             ?: return problemReporter.problemLastMessage(invitation.threadId, "goal-code-unknown")
         val message = buildRequestCredentialMessage(credentialScheme, invitation.id)
@@ -174,7 +175,7 @@ class IssueCredentialProtocol(
         val serviceEndpoint = invitation.body.services?.let {
             if (it.isNotEmpty()) it[0].serviceEndpoint else null
         }
-        return InternalNextMessage.SendAndWrap(message, senderKeyId, serviceEndpoint)
+        return InternalNextMessage.SendAndWrap(message, senderKey = senderKey, endpoint = serviceEndpoint)
             .also { this.threadId = message.threadId }
             .also { this.state = State.REQUEST_CREDENTIAL_SENT }
     }
@@ -213,7 +214,7 @@ class IssueCredentialProtocol(
         )
     }
 
-    private suspend fun issueCredential(lastMessage: RequestCredential, senderKeyId: String): InternalNextMessage {
+    private suspend fun issueCredential(lastMessage: RequestCredential, senderKey: JsonWebKey): InternalNextMessage {
         val lastJwmAttachment = lastMessage.attachments?.firstOrNull()
             ?: return problemReporter.problemLastMessage(lastMessage.threadId, "attachments-missing")
         val requestCredentialAttachment = lastJwmAttachment.decodeString()?.let {
@@ -222,7 +223,7 @@ class IssueCredentialProtocol(
         val uri = requestCredentialAttachment.credentialManifest.credential.schema.uri
 
         //default pupilID use case: binding key outside JWM, no subject in JWM. set subject to override
-        val subjectKeyId = requestCredentialAttachment.credentialManifest.subject ?: senderKeyId
+        val subjectKeyId = requestCredentialAttachment.credentialManifest.subject ?: senderKey.getIdentifier()
 
         val requestedAttributeType = AttributeIndex.getTypeOfAttributeForSchemaUri(uri)
             ?: return problemReporter.problemLastMessage(lastMessage.threadId, "requested-attributes-empty")
@@ -271,7 +272,7 @@ class IssueCredentialProtocol(
             threadId = lastMessage.threadId!!, //is allowed to fail horribly
             attachments = (fulfillmentAttachments + binaryAttachments).toTypedArray()
         )
-        return InternalNextMessage.SendAndWrap(message, senderKeyId)
+        return InternalNextMessage.SendAndWrap(message, senderKey = senderKey)
             .also { this.threadId = message.threadId }
             .also { this.state = State.FINISHED }
     }